Adds decimal to binary converter.
This is not a very complicated task, but is allows for quick conversion. Uses a string
to represent the binary number.

It also sets up the project to allow for multiple different types of mini-apps to be run.
This way I can add all my little challenges to the same project and test them.